# stream: optimize JsonFraming